+++ R/exclude_nca.R
@@ -6,6 +6,8 @@
 #'   `PKNCA.options("max.aucinf.pext")` if not provided).
 #' @param min.hl.r.squared The minimum acceptable r-squared value for half-life
 #'   (uses `PKNCA.options("min.hl.r.squared")` if not provided).
+#' @param min.hl.adj.r.squared The minimum acceptable adjusted r-squared for half-life
+#'   (uses 0.9 if not provided).
 #' @examples
 #' my_conc <- PKNCAconc(data.frame(conc=1.1^(3:0),
 #'                                 time=0:3,
@@ -26,77 +28,57 @@ NULL
 #' @describeIn exclude_nca Exclude based on span.ratio
 #' @export
 exclude_nca_span.ratio <- function(min.span.ratio) {
-  affected_parameters <- get.parameter.deps("half.life")
   missing_min.span.ratio <- missing(min.span.ratio)
-  function(x, ...) {
-    if (missing_min.span.ratio) {
-      min.span.ratio <- PKNCA.options("min.span.ratio")
-    }
-    ret <- rep(NA_character_, nrow(x))
-    if (!is.na(min.span.ratio)) {
-      idx_span.ratio <- which(x$PPTESTCD %in% "span.ratio")
-      if (length(idx_span.ratio) == 0) {
-        # Do nothing, it wasn't calculated
-      } else if (length(idx_span.ratio) == 1) {
-        current_span.ratio <- x$PPORRES[idx_span.ratio]
-        drop_span_ratio <-
-          !is.na(current_span.ratio) &
-          current_span.ratio < min.span.ratio
-        if (drop_span_ratio) {
-          ret[x$PPTESTCD %in% affected_parameters] <-
-            sprintf("Span ratio < %g", min.span.ratio)
-        }
-      } else if (length(idx_span.ratio) > 1) { # nocov
-        stop("Should not see more than one span.ratio (please report this as a bug)") # nocov
-      }
-    }
-    ret
+  if (missing_min.span.ratio) {
+    min.span.ratio <- PKNCA.options("min.span.ratio")
   }
+  exclude_nca_by_param(
+    parameter = "span.ratio",
+    min_thr = min.span.ratio,
+    affected_parameters = get.parameter.deps("half.life")
+  )
 }
 
-#' @describeIn exclude_nca Exclude based on AUC percent extrapolated (both
-#'   observed and predicted)
+#' @describeIn exclude_nca Exclude based on AUC percent extrapolated (both observed and predicted)
 #' @export
-exclude_nca_max.aucinf.pext <-  function(max.aucinf.pext) {
-  affected_parameters <-
-    list(obs=get.parameter.deps("aucinf.obs"),
-         pred=get.parameter.deps("aucinf.pred"))
+exclude_nca_max.aucinf.pext <- function(max.aucinf.pext) {
   missing_max.aucinf.pext <- missing(max.aucinf.pext)
+  if (missing_max.aucinf.pext) {
+    max.aucinf.pext <- PKNCA.options("max.aucinf.pext")
+  }
+  # Exclude for both obs and pred
   function(x, ...) {
-    if (missing_max.aucinf.pext) {
-      max.aucinf.pext <- PKNCA.options("max.aucinf.pext")
-    }
-    ret <- rep(NA_character_, nrow(x))
-    if (!is.na(max.aucinf.pext)) {
-      for (ext_type in c("obs", "pred")) {
-        idx_pext <- which(x$PPTESTCD %in% paste0("aucpext.", ext_type))
-        if (length(idx_pext) == 0) {
-          # Do nothing, it wasn't calculated
-        } else if (length(idx_pext) == 1) {
-          current_pext <- x$PPORRES[idx_pext]
-          drop_pext <-
-            !is.na(current_pext) &
-            current_pext > max.aucinf.pext
-          if (drop_pext) {
-            ret[x$PPTESTCD %in% affected_parameters[[ext_type]]] <-
-              sprintf("AUC percent extrapolated > %g", max.aucinf.pext)
-          }
-        } else if (length(idx_pext) > 1) { # nocov
-          stop("Should not see more than one aucpext.", ext_type, " (please report this as a bug)") # nocov
-        }
-      }
-    }
+    res_obs <- exclude_nca_by_param(
+      parameter = "aucpext.obs",
+      max_thr = max.aucinf.pext,
+      affected_parameters = get.parameter.deps("aucinf.obs")
+    )(x, ...)
+    res_pred <- exclude_nca_by_param(
+      parameter = "aucpext.pred",
+      max_thr = max.aucinf.pext,
+      affected_parameters = get.parameter.deps("aucinf.pred")
+    )(x, ...)
+    # Combine results, prioritizing non-NA from either
+    is.obs <- grepl(paste0("aucpext.obs > ", max.aucinf.pext), res_obs)
+    is.pred <- grepl(paste0("aucpext.pred > ", max.aucinf.pext), res_pred)
+    ret <- ifelse(
+      is.obs | is.pred,
+      gsub(
+        pattern = paste0("aucpext...+ > ", max.aucinf.pext),
+        replacement = paste0("aucpext > ", max.aucinf.pext),
+        x = dplyr::coalesce(res_obs, res_pred)
+      ),
+      res_obs
+    )
     ret
   }
 }
 
-#' @describeIn exclude_nca Exclude AUC measurements based on count of
-#'   concentrations measured and not below the lower limit of quantification
+#' @describeIn exclude_nca Exclude AUC measurements based on count of concentrations measured and not below the lower limit of quantification
 #' @param min_count Minimum number of measured concentrations
-#' @param exclude_param_pattern Character vector of regular expression patterns
-#'   to exclude
+#' @param exclude_param_pattern Character vector of regular expression patterns to exclude
 #' @export
-exclude_nca_conc_count_measured <-  function(min_count, exclude_param_pattern = c("^aucall", "^aucinf", "^aucint", "^auciv", "^auclast", "^aumc", "^sparse_auc")) {
+exclude_nca_count_conc_measured <- function(min_count, exclude_param_pattern = c("^aucall", "^aucinf", "^aucint", "^auciv", "^auclast", "^aumc", "^sparse_auc")) {
   all_parameters <- names(PKNCA::get.interval.cols())
   affected_parameters_base <-
     sort(unique(unlist(
@@ -114,55 +96,147 @@ exclude_nca_conc_count_measured <-  function(min_count, exclude_param_pattern = 
         FUN = get.parameter.deps
       )
     )))
-  force(min_count)
-  function(x, ...) {
-    ret <- rep(NA_character_, nrow(x))
-    if (!is.na(min_count)) {
-      idx_count <- which(x$PPTESTCD %in% "count_conc_measured")
-      if (length(idx_count) == 0) {
-        # Do nothing, it wasn't calculated
-      } else if (length(idx_count) == 1) {
-        current_count <- x$PPORRES[idx_count]
-        drop_count <-
-          !is.na(current_count) &
-          current_count < min_count
-        if (drop_count) {
-          ret[x$PPTESTCD %in% affected_parameters] <-
-            sprintf("Number of measured concentrations is < %g", min_count)
-        }
-      } else { # nocov
-        stop("Should not see more than one count_conc_measured (please report this as a bug)") # nocov
-      }
-    }
-    ret
-  }
+  exclude_nca_by_param(
+    parameter = "count_conc_measured",
+    min_thr = min_count,
+    affected_parameters = affected_parameters
+  )
 }
 
 #' @describeIn exclude_nca Exclude based on half-life r-squared
 #' @export
 exclude_nca_min.hl.r.squared <- function(min.hl.r.squared) {
-  affected_parameters <- get.parameter.deps("half.life")
   missing_min.hl.r.squared <- missing(min.hl.r.squared)
-  function(x, ...) {
-    if (missing_min.hl.r.squared) {
-      min.hl.r.squared <- PKNCA.options("min.hl.r.squared")
+  if (missing_min.hl.r.squared) {
+    min.hl.r.squared <- PKNCA.options("min.hl.r.squared")
+  }
+  exclude_nca_by_param(
+    parameter = "r.squared",
+    min_thr = min.hl.r.squared,
+    affected_parameters = get.parameter.deps("half.life")
+  )
+}
+
+#' @describeIn exclude_nca Exclude based on half-life adjusted r-squared
+#' @export
+exclude_nca_min.hl.adj.r.squared <- function(min.hl.adj.r.squared = 0.9) {
+  exclude_nca_by_param(
+    parameter = "adj.r.squared",
+    min_thr = min.hl.adj.r.squared,
+    affected_parameters = get.parameter.deps("half.life")
+  )
+}
+
+#' @describeIn exclude_nca Exclude based on implausibly early Tmax (often used for extravascular dosing with a Tmax value of 0)
+#' @param tmax_early The time for Tmax which is considered too early to be a valid NCA result
+#' @export
+exclude_nca_tmax_early <- function(tmax_early = 0) {
+    
+    exc_fun <- exclude_nca_by_param(
+      parameter = "tmax",
+      min_thr = tmax_early,
+      affected_parameters = names(get.interval.cols())
+    )
+
+    function(x, ...) {
+    # Get the exclusion messages
+    ret <- exc_fun(x, ...)
+    # Add the special annotation for tmax_early cases (if not already annotated)
+    tmax_cases <- grepl(pattern = "^tmax < ", x = ret)
+    tmax_already_annotated <- grepl("(likely missed dose, insufficient PK samples, or PK sample swap)", x = ret, fixed = TRUE)
+    ret <- ifelse(
+        test = tmax_cases & !tmax_already_annotated & !is.na(ret),
+        yes = gsub(
+          pattern = paste0("^tmax < ", tmax_early),
+          replacement = paste0("tmax < ", tmax_early, " (likely missed dose, insufficient PK samples, or PK sample swap)"),
+          x = ret
+        ),
+        no = ret
+      )
+    ret
     }
+}
+
+#' @describeIn exclude_nca Exclude based on implausibly early Tmax (special case for `tmax_early = 0`)
+#' @export
+exclude_nca_tmax_0 <- function() {
+  exc_fun <- exclude_nca_tmax_early(1e-99)
+  function(x, ...) {
+    ret <- exc_fun(x, ...)
+
+    # Replace the messages
+    ret <- gsub(
+      pattern = "^tmax < 1e-99",
+      replacement = "tmax <= 0",
+      x = ret
+    )
+    ret
+  }
+}
+
+
+#' Exclude NCA Results Based on Parameter Thresholds
+#'
+#' Exclude rows from NCA results based on specified thresholds for a given parameter.
+#' This function allows users to define minimum and/or maximum acceptable values
+#' for a parameter and excludes rows that fall outside these thresholds.
+#'
+#' @param parameter The name of the PKNCA parameter to evaluate (e.g., "span.ratio").
+#' @param min_thr The minimum acceptable value for the parameter. If not provided, is not applied.
+#' @param max_thr The maximum acceptable value for the parameter. If not provided, is not applied.
+#' @param affected_parameters Character vector of PKNCA parameters that will be marked as excluded.
+#'                            By default is the defined parameter.
+#' @returns A function that can be used with `PKNCA::exclude` to mark through the 'exclude'  column
+#'          the rows in the PKNCA results based on the specified thresholds for a parameter.
+#' @examples
+#' # Example dataset
+#' my_data <- PKNCA::PKNCAdata(
+#'   PKNCA::PKNCAconc(data.frame(conc = 5:1,
+#'                               time = 0:4,
+#'                               subject = 1),
+#'                    conc ~ time | subject),
+#'   PKNCA::PKNCAdose(data.frame(subject = 1, dose = 100, time = 0),
+#'                    dose ~ time | subject)
+#' )
+#' my_result <- PKNCA::pk.nca(my_data)
+#'
+#' # Exclude rows where span.ratio is less than 2
+#' excluded_result <- PKNCA::exclude(
+#'   my_result,
+#'   FUN = exclude_nca_by_param("span.ratio", min_thr = 2)
+#' )
+#' as.data.frame(excluded_result)
+#'
+#' @export
+
+exclude_nca_by_param <- function(
+  parameter,
+  min_thr = NULL,
+  max_thr = NULL,
+  affected_parameters = parameter
+) {
+  # Check that defined thresholds are single numeric objects
+  checkmate::expect_number(min_thr, finite = TRUE, null.ok = TRUE)
+  checkmate::expect_number(max_thr, finite = TRUE, null.ok = TRUE)
+
+  if (isTRUE(min_thr > max_thr))
+    stop("if both defined min_thr must be less than max_thr")
+
+  function(x, ...) {
     ret <- rep(NA_character_, nrow(x))
-    if (!is.na(min.hl.r.squared)) {
-      idx_r.squared <- which(x$PPTESTCD %in% "r.squared")
-      if (length(idx_r.squared) == 0) {
-        # Do nothing, it wasn't calculated
-      } else if (length(idx_r.squared) == 1) {
-        current_r.squared <- x$PPORRES[idx_r.squared]
-        drop_r.squared <-
-          !is.na(current_r.squared) &
-          current_r.squared < min.hl.r.squared
-        if (drop_r.squared) {
-          ret[x$PPTESTCD %in% affected_parameters] <-
-            sprintf("Half-life r-squared < %g", min.hl.r.squared)
-        }
-      } else if (length(idx_r.squared) > 1) { # nocov
-        stop("Should not see more than one r.squared (please report this as a bug)") # nocov
+    idx_param <- which(x$PPTESTCD == parameter)
+    idx_aff_params <- which(x$PPTESTCD %in% affected_parameters)
+
+    if (length(idx_param) > 1)
+      stop("Should not see more than one ", parameter, " (please report this as a bug)")
+
+    if (length(idx_param) == 1 && !is.na(x$PPORRES[idx_param]) && length(idx_aff_params) > 0) {
+      current_value <- x$PPORRES[idx_param]
+      pretty_name <- parameter # Pretty name did not convince me for some parameters (e.g, "r.squared")
+      if (isTRUE(current_value < min_thr)) {
+        ret[idx_aff_params] <- sprintf("%s < %g", pretty_name, min_thr)
+      } else if (isTRUE(current_value > max_thr)) {
+        ret[idx_aff_params] <- sprintf("%s > %g", pretty_name, max_thr)
       }
     }
     ret