+++ tests/testthat/test-class-PKNCAdata.R
@@ -358,45 +358,31 @@ test_that("intervals may be a tibble", {
   )
 })
 
-test_that("getGroups works", {
-  # Check that it works with grouping [contains only the grouping column(s)]
-  o_conc_group <- PKNCAconc(as.data.frame(datasets::Theoph), conc~Time|Subject)
-  data_group <- as.data.frame(datasets::Theoph)
-  expected_group <- data.frame(Subject = data_group$Subject)
-
-  expect_equal(getGroups.PKNCAdata(o_conc_group), expected_group)
-
-  # Check that it works without groupings as expected [empty]
-  o_conc_nongroup <- PKNCAconc(as.data.frame(datasets::Theoph)[datasets::Theoph$Subject == 1,], conc~Time)
-
-  expect_equal(names(getGroups.PKNCAdata(o_conc_nongroup)), character(0))
-})
-
 test_that("PKNCAdata units (#336)", {
   # Typical use
   d_conc <- data.frame(conc = 1, time = 0, concu_x = "A", timeu_x = "B", amountu_x = "C")
-  d_dose <- data.frame(dose = 1, time = 0, doseu_x = "B")
+  d_dose <- data.frame(dose = 1, time = 0, doseu_x = "D")
 
-  o_conc <- PKNCAconc(data = d_conc, conc~time, concu = "concu_x")
+  o_conc <- PKNCAconc(data = d_conc, conc~time, concu = "concu_x", timeu = "timeu_x")
   o_dose <- PKNCAdose(data = d_dose, dose~time, doseu = "doseu_x")
   o_data <- PKNCAdata(o_conc, o_dose)
   expect_equal(
     o_data$units,
-    pknca_units_table(concu = "A", doseu = "B")
+    pknca_units_table(concu = "A", doseu = "D", timeu = "B")
   )
   suppressWarnings(o_nca <- pk.nca(o_data))
   expect_true("Cmax (A)" %in% names(summary(o_nca)))
 
   # NA unit values are ignored
   d_conc <- data.frame(conc = 1, time = 0:1, concu_x = c("A", NA), timeu_x = "B", amountu_x = "C")
-  d_dose <- data.frame(dose = 1, time = 0, doseu_x = "B")
+  d_dose <- data.frame(dose = 1, time = 0, doseu_x = "D")
 
-  o_conc <- PKNCAconc(data = d_conc, conc~time, concu = "concu_x")
+  o_conc <- PKNCAconc(data = d_conc, conc~time, concu = "concu_x", timeu = "timeu_x")
   o_dose <- PKNCAdose(data = d_dose, dose~time, doseu = "doseu_x")
   o_data <- PKNCAdata(o_conc, o_dose)
   expect_equal(
     o_data$units,
-    pknca_units_table(concu = "A", doseu = "B")
+    pknca_units_table(concu = "A", doseu = "D", timeu = "B")
   )
   suppressWarnings(o_nca <- pk.nca(o_data))
   expect_true("Cmax (A)" %in% names(summary(o_nca)))
@@ -411,4 +397,33 @@ test_that("PKNCAdata units (#336)", {
     PKNCAdata(o_conc, o_dose),
     regexp = "Only one unit may be provided at a time: A, C"
   )
+})
+
+test_that("getGroups works", {
+  # Check that it works with grouping [contains only the grouping column(s)]
+  o_conc_group <- PKNCAconc(as.data.frame(datasets::Theoph), conc~Time|Subject)
+  data_group <- as.data.frame(datasets::Theoph)
+  expected_group <- data.frame(Subject = data_group$Subject)
+  o_data_group <- PKNCAdata(o_conc_group, intervals = data.frame(start = 0, end = 1, cmax = TRUE))
+  expect_equal(getGroups(o_data_group), expected_group)
+
+  # Check that it works without groupings as expected [empty]
+  o_conc_nongroup <- PKNCAconc(as.data.frame(datasets::Theoph)[datasets::Theoph$Subject == 1,], conc~Time)
+  o_data_nogroup <- PKNCAdata(o_conc_nongroup, intervals = data.frame(start = 0, end = 1, cmax = TRUE))
+
+  # It should be an empty data.frame with 11 rows
+  expect_equal(getGroups(o_data_nogroup), data.frame(A = 1:11)[, -1])
+})
+
+test_that("group_vars.PKNCAdata", {
+  o_conc_group <- PKNCAconc(as.data.frame(datasets::Theoph), conc~Time|Subject)
+  o_data_group <- PKNCAdata(o_conc_group, intervals = data.frame(start = 0, end = 1, cmax = TRUE))
+
+  expect_equal(dplyr::group_vars(o_data_group), "Subject")
+
+  # Check that it works without groupings as expected [empty]
+  o_conc_nongroup <- PKNCAconc(as.data.frame(datasets::Theoph)[datasets::Theoph$Subject == 1,], conc~Time)
+  o_data_nogroup <- PKNCAdata(o_conc_nongroup, intervals = data.frame(start = 0, end = 1, cmax = TRUE))
+
+  expect_equal(dplyr::group_vars(o_data_nogroup), character(0))
 })