+++ NEWS.md
@@ -4,7 +4,56 @@ will continue until then.  These will be especially noticeable around
 the inclusion of IV NCA parameters and additional specifications of
 the dosing including dose amount and route.
 
-# PKNCA 0.11.0.9000
+# Development version
+
+## Breaking changes
+
+* Both include and excluding half-life points may not be done for the same interval (#406)
+
+## Bugs fixed
+
+* `get_halflife_points()` now correctly accounts for start time != 0 and sets
+  times outside of any interval to `NA` (#470)
+
+## New features
+
+* `pk.calc.half.life()` now returns also `lambda.z.corrxy`, the correlation between
+  the time and the log-concentration of the lambda z points.
+* `get_halflife_points()` can now be used on PKNCAdata objects to see which points
+  would be used for half-life calculation (#476)
+* New excretion parameters: `volpk` (total urine volume for an interval) and
+  dose-normalized renal clearance parameters: `clr.last.dn`, `clr.obs.dn`,
+  `clr.pred.dn` (#433)
+* `PKNCA.set.summary(reset = TRUE)` warns that it may break the use of
+  `summary()` (#477)
+
+* New post-processing functions to normalize PKNCA result parameters based on any column in PKNCAconc data.frame (`normalize_by_col()`) or by using a custom normalization table (`normalize()`)
+* New excretion rate parameters: `ermax` (Maximum excretion rate), `ertmax` (Midpoint time
+  of maximum excretion rate) and `ertlst` (Time of last excretion rate measurement) (#433)
+
+# PKNCA 0.12.1
+
+## Minor changes (unlikely to affect PKNCA use)
+
+* Units for fraction excretion parameter (fe) are now accurately captured as
+  amount/dose units rather than "fraction" (#426)
+* `get_halflife_points` will ignore points after `lambda.z.time.last`, instead
+  of `tlast` (#448)
+* `lambda.z` calculations will now only consider time points that occur after
+  the end of the latest dose administration (#139)
+* `aucint.inf.pred` is `NA` when half-life is not estimable (#450)
+
+## New features
+
+* PKNCA now has a debugging mode to support troubleshooting; it is not intended
+  for production use. Debugging mode can be enabled using
+  `PKNCA.options(debug = TRUE)`.
+* It is now possible to update an existing analysis when data changes but other
+  NCA settings stay the same (fix #417)
+* New assertion functions were created to ensure that an object is the correct
+  type (fix #328)
+
+# PKNCA 0.12.0
 
 ## Breaking changes
 
@@ -15,6 +64,7 @@ the dosing including dose amount and route.
   calculate `c0` and does not raise an error when `is.na(c0)` (#353).
 * Manual calculation of half.life no longer allows negative half-live values
   (#373).
+* pk.calc.half.life() now returns also lambda.z.time.last, the last time point used for terminal slope estimation.
 
 ## New Features
 
@@ -24,13 +74,23 @@ the dosing including dose amount and route.
   names in the `conc.blq` argument (`before.tmax`,`after.tmax`)
 * A new parameter `count_conc_measured` was added to enable quality checks,
   typically on AUC measurements. An associated exclusion function,
-  `exclude_nca_conc_count_measured()` was also added.
+  `exclude_nca_count_conc_measured()` was also added.
 * The `PKNCAconc()` arguments of `include_half.life` and `exclude_half.life` now
   allow `NA` values. If all values are `NA`, then no inclusion or exclusion is
   applied (the interval is treated as-is, like the argument had not been given).
   If some values are `NA` for the interval, those are treated as `FALSE`.
+* `group_vars()` methods were added for `PKNCAdata` and `PKNCAresults` objects.
+* If intervals have attributes on the columns, there will no longer be an error
+  during parameter calculation, and the attributes are preserved (#381)
+* When adding units, if some but not all units are provided, then an error will
+  be raised. This error can be converted to a warning using the option
+  `allow_partial_missing_units = TRUE`. (#398)
+* A new function `get_halflife_points()` lets users know which points were used
+  for half-life calculation. (#387)
+* A new function `exclude_nca_min.hl.adj.r.squared()` to allow exclusion of
+  half-life results based on a minimum adjusted r-squared threshold.
 
-# Minor changes (unlikely to affect PKNCA use)
+## Minor changes (unlikely to affect PKNCA use)
 
 * PKNCA will now verify the `intervals` data.frame when creating PKNCAdata. The
   checking includes confirming intended column naming and ensuring the correct
@@ -42,6 +102,7 @@ the dosing including dose amount and route.
 * Removed native pipes (`|>`) so that PKNCA will work with older versions of R
   (#304).
 * Missing dosing times to `pk.calc.c0()` will not cause an error (#344)
+* `getGroups()` includes the `end` column when applied to a `PKNCAresults` object (#419).
 
 # PKNCA 0.11.0
 
@@ -282,7 +343,7 @@ the dosing including dose amount and route.
 # PKNCA 0.9.0
 
 * Breaking Change: `plot.PKNCAconc()` was moved to the pknca.reporting package
-  (https://github.com/billdenney/pknca.reporting)
+  (https://github.com/humanpred/pknca.reporting)
 * Breaking Change: `summary.PKNCAresults()` now provides a caption
   including the summary method for each parameter.  If you change
   summary functions using `PKNCA.set.summary()`, you must now use the